+++ src/windows/ui.rs
@@ -1,0 +1,2 @@
+// Derive from:
+// https://github.com/microsoft/Windows-classic-samples/blob/master/Samples/Win7Samples/winui/shell/appshellintegration/NotificationIcon/NotificationIcon.cpp
